+++ bodo/tests/utils.py
@@ -695,7 +695,7 @@ def _test_equal(
     elif pd.api.types.is_array_like(py_out) and pd.api.types.is_extension_array_dtype(
         py_out
     ):
-        # bodo currently returns np object array instead of pd StringArray
+        # bodo currently returns object array instead of pd StringArray
         if not pd.api.types.is_extension_array_dtype(bodo_out):
             bodo_out = pd.array(bodo_out)
         if sort_output:
@@ -1134,7 +1134,7 @@ def convert_non_pandas_columns(df):
         e_col = df[e_col_name]
 
         def is_ok(val):
-            if isinstance(val, np.float):
+            if isinstance(val, float):
                 return not np.isnan(val)
             if val == None:
                 return False
